module structure inside state key file
