+++ crates/core/component/funding/src/component/liquidity_tournament/mod.rs
@@ -1,0 +1,1 @@
+pub mod nullifier;
